requirements.txt
databutton
ta
ccxt
pandas 
streamlit-option-menu 
discord-webhook
plotly
streamlit-autorefresh